requirements.txt
python-dotenv
psycopg2-binary
pandas
numpy